+++ src/CognitiveServices/CognitiveServices.Test/ScenarioTests/CognitiveServicesAccountTests.cs
@@ -211,5 +211,12 @@ namespace CognitiveServices.Test.ScenarioTests
         {
             TestController.NewInstance.RunPsTest(traceInterceptor, "Test-PublicNetworkAccess");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestCapabilities()
+        {
+            TestController.NewInstance.RunPsTest(traceInterceptor, "Test-Capabilities");
+        }
     }
 }